requirements_ingest.txt
# Connectivity & API
httpx==0.28.1
google-cloud-pubsub==2.27.1

# Logic & Reliability
tenacity==9.0.0
python-dotenv==1.01
click==8.3.1
